refactor: migrate schema to core module, drop registry_resolver
- Move config/ and features/ from schema/ to core/
- Remove unused registry_resolver.rs and GitignoreScope trait
- Suppress dead code warnings, clean up unused exports and imports
