+++ src/cli/mod.rs
@@ -8,5 +8,5 @@ mod skills;
 pub(crate) mod ui;
 mod undeploy;
 
-pub(crate) use options::{InitOptions, get_options};
+pub(crate) use options::get_options;
 pub(crate) use runner::run;
