+++ src/constants/schema.rs
@@ -1,4 +1,6 @@
 pub(crate) const CONFIG_SCHEMA: &str = "https://dotagents.soorya-u.dev/schemas/config.schema.json";
+#[allow(dead_code)]
 pub(crate) const MCP_SCHEMA: &str = "https://dotagents.soorya-u.dev/json/schemas/mcp.schema.json";
+#[allow(dead_code)]
 pub(crate) const PROVIDER_SCHEMA: &str =
     "https://dotagents.soorya-u.dev/json/schemas/provider.schema.json";
