+++ src/core/config/local.rs
@@ -3,16 +3,16 @@ use std::collections::{HashMap, HashSet};
 use super::common::{PackageRunner, Providers, Targets};
 use super::traits::TomlConfig;
 use crate::constants::schema::CONFIG_SCHEMA;
-use crate::schema::features::Feature;
+use crate::core::features::Feature;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(rename_all = "kebab-case")]
-pub struct GlobalConfig {
+pub struct LocalConfig {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub schema: Option<String>,
-    #[serde(default)]
-    pub features: HashSet<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub features: Option<HashSet<String>>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub targets: Option<Targets>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -23,11 +23,11 @@ pub struct GlobalConfig {
     pub package_runner: Option<PackageRunner>,
 }
 
-impl GlobalConfig {
+impl LocalConfig {
     pub fn new() -> Self {
         Self {
             schema: Some(CONFIG_SCHEMA.into()),
-            features: HashSet::new(),
+            features: None,
             targets: None,
             providers: None,
             variables: None,
@@ -35,46 +35,71 @@ impl GlobalConfig {
         }
     }
 
-    pub fn with_features(features: HashSet<String>, targets: Targets) -> Self {
+    #[allow(dead_code)]
+    pub fn with_features(features: HashSet<String>) -> Self {
         Self {
             schema: Some(CONFIG_SCHEMA.into()),
-            features,
-            targets: Some(targets),
+            features: Some(features),
+            targets: None,
             providers: None,
             variables: None,
             package_runner: None,
         }
     }
 
+    #[allow(dead_code)]
+    pub fn with_providers(providers: Providers) -> Self {
+        Self {
+            schema: Some(CONFIG_SCHEMA.into()),
+            features: None,
+            targets: None,
+            providers: Some(providers),
+            variables: None,
+            package_runner: None,
+        }
+    }
+
     pub fn validate(&self) -> anyhow::Result<()> {
-        for feature in &self.features {
-            if Feature::from_str(feature).is_none() {
-                anyhow::bail!(
-                    "Invalid feature: {}. Valid features are: {}",
-                    feature,
-                    Feature::all_names().join(", ")
-                );
+        // Check that features are valid if present
+        if let Some(features) = &self.features {
+            for feature in features {
+                if Feature::from_str(feature).is_none() {
+                    anyhow::bail!(
+                        "Invalid feature: {}. Valid features are: {}",
+                        feature,
+                        Feature::all_names().join(", ")
+                    );
+                }
             }
         }
 
         Ok(())
     }
+
+    #[allow(dead_code)]
+    pub fn is_empty(&self) -> bool {
+        self.schema.is_none()
+            && self.features.is_none()
+            && self.targets.is_none()
+            && self.providers.is_none()
+            && self.package_runner.is_none()
+    }
 }
 
-impl Default for GlobalConfig {
+impl Default for LocalConfig {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl TomlConfig for GlobalConfig {}
+impl TomlConfig for LocalConfig {}
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
-    fn package_runner_field_deserialises_in_global_config() {
+    fn package_runner_field_deserialises_in_local_config() {
         // all four runner values parse correctly from TOML
         for (toml_val, expected) in [
             ("npm", PackageRunner::Npm),
@@ -83,22 +108,15 @@ mod tests {
             ("bun", PackageRunner::Bun),
         ] {
             let toml = format!("package-runner = \"{toml_val}\"\n");
-            let config: GlobalConfig = toml::from_str(&toml).unwrap();
+            let config: LocalConfig = toml::from_str(&toml).unwrap();
             assert_eq!(config.package_runner, Some(expected));
         }
     }
 
     #[test]
-    fn package_runner_absent_yields_none_in_global_config() {
+    fn package_runner_absent_yields_none_in_local_config() {
         // omitting the field deserialises to None
-        let config: GlobalConfig = toml::from_str("features = []\n").unwrap();
+        let config: LocalConfig = toml::from_str("").unwrap();
         assert_eq!(config.package_runner, None);
-    }
-
-    #[test]
-    fn invalid_package_runner_value_fails_deserialisation() {
-        // an unrecognised runner value should fail with a serde error
-        let result: Result<GlobalConfig, _> = toml::from_str("package-runner = \"cargo\"\n");
-        assert!(result.is_err());
     }
 }
