+++ src/core/config/traits.rs
@@ -2,6 +2,7 @@ use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
 
 pub trait TomlConfig: Serialize + for<'de> Deserialize<'de> {
+    #[allow(dead_code)]
     fn to_toml(&self) -> Result<String> {
         toml::to_string_pretty(self).context("Failed to serialize configuration to TOML")
     }
