+++ src/core/features/traits.rs
@@ -1,8 +1,7 @@
-use anyhow::{Result, anyhow};
+use anyhow::Result;
 use serde_json::Value;
 
 use crate::templates::{RenderType, Templater};
-use crate::utils::gitignore::GitignoreScope;
 
 pub trait FeatureTrait: Sized {
     fn from_string(value: &str) -> Result<Self>;
@@ -20,14 +19,8 @@ pub trait FeatureTrait: Sized {
         None
     }
 
-    /// Returns the name variable for target path interpolation.
     /// Must be overridden by any feature that returns `Some` from `get_file_name`.
     fn get_name_variable(&self, _filename: &str) -> Result<Option<Value>> {
         Ok(None)
-    }
-
-    /// Returns how this feature's deployed paths should be represented in .gitignore.
-    fn gitignore_scope(&self) -> GitignoreScope {
-        GitignoreScope::File
     }
 }
