+++ src/core/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod config;
+pub(crate) mod features;
