+++ src/prelude.rs
@@ -1,3 +1,3 @@
 pub(crate) use anyhow::{Context, Result, anyhow, bail};
 
-pub(crate) use crate::{debug, error, info, step, success, trace, warn};
+pub(crate) use crate::{debug, error, info, success, warn};
