+++ src/schema/mod.rs
@@ -1,3 +1,1 @@
-pub(crate) mod config;
-pub(crate) mod features;
 pub(crate) mod registry;
